backend/requirements.txt
# ThoughtField — backend/requirements.txt
# Install with: pip install -r requirements.txt

# ---- Web framework ----
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9

# ---- Data validation ----
pydantic==2.7.1
pydantic-settings==2.2.1

# ---- LLM ----
openai==1.30.1

# ---- Vector DB (agent memory) ----
chromadb==0.5.0

# ---- Cache / pub-sub (live state streaming) ----
redis==5.0.4
redis[asyncio]

# ---- Utilities ----
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0        # used internally by openai SDK

# ---- Dev / testing ----
pytest==8.2.0
pytest-asyncio==0.23.6